refactor: flatten subworkflow adapter package
